refactor(worker): isolate workspace handoff helpers
